+++ tests/document_controller/test_action_save_document.py
@@ -27,14 +27,14 @@ from mtg_proxy_printer.sqlite_helpers import open_database, create_in_memory_dat
 from mtg_proxy_printer.units_and_sizes import unit_registry, UnitT, CardSizes
 from mtg_proxy_printer.model.carddb import CheckCard
 from mtg_proxy_printer.model.document import Document
-from mtg_proxy_printer.model.document_loader import DocumentLoader, CardType
+from mtg_proxy_printer.model.document_loader import CardType
 from mtg_proxy_printer.model.imagedb_files import ImageKey
 from mtg_proxy_printer.document_controller.card_actions import ActionAddCard
 from mtg_proxy_printer.document_controller.edit_document_settings import ActionEditDocumentSettings
 from mtg_proxy_printer.document_controller.save_document import ActionSaveDocument
 
 from tests.test_document import document_custom_layout
-from tests.helpers import close_to_
+from tests.helpers import close_to_, quantity_close_to
 
 ItemDataRole = Qt.ItemDataRole
 mm: UnitT = unit_registry.mm
@@ -45,7 +45,7 @@ def validate_qt_model_signal_parameter(
     return not parent.isValid() and first == expected_first and last == expected_last
 
 
-@pytest.mark.parametrize("source_version", [2, 3, 4, 5])
+@pytest.mark.parametrize("source_version", [2, 3, 4, 5, 6])
 def test_save_migration(tmp_path: Path, document: Document, source_version: int):
     """Tests migration of existing saves to the newest schema revision on save."""
     card = document.card_db.get_card_with_scryfall_id("0000579f-7b35-4ed3-b44c-db2a538066fe", True)
@@ -76,8 +76,7 @@ def test_save_as_saves_regular_card(tmp_path: Path, document: Document, is_front
     save_file = tmp_path/"test.mtgproxies"
     action = ActionSaveDocument(save_file)
     action.apply(document)
-    with open_database(
-            save_file, "document-v6", DocumentLoader.MIN_SUPPORTED_SQLITE_VERSION, False) as con:
+    with open_database(save_file, "document-v6", False) as con:
         content = con.execute("SELECT page, slot, scryfall_id, is_front, type FROM Card").fetchall()
     del con
     assert_that(
@@ -96,8 +95,7 @@ def test_save_as_saves_check_card(tmp_path: Path, document: Document):
     save_file = tmp_path / "test.mtgproxies"
     action = ActionSaveDocument(save_file)
     action.apply(document)
-    with open_database(
-            save_file, "document-v6", DocumentLoader.MIN_SUPPORTED_SQLITE_VERSION, False) as con:
+    with open_database(save_file, "document-v6") as con:
         content = con.execute("SELECT page, slot, scryfall_id, is_front, type FROM Card").fetchall()
     del con
     assert_that(
@@ -130,14 +128,14 @@ def test_subsequent_save_updates_settings(tmp_path: Path, qtbot: QtBot, document
                            check_params_cbs=[lambda value: value, lambda value: not value]):
         document_custom_layout.loader.load_document(save_file)
     assert_that(
-        document_custom_layout.page_layout.page_height.to(mm).magnitude,
-        is_(close_to_(1000)))
+        document_custom_layout.page_layout.page_height,
+        is_(quantity_close_to(1000*mm)))
 
 
 def _create_save_file(temp_path: Path, source_version: int):
     """Creates an empty document save file at the given path and using the given schema version."""
     save_file_path = temp_path/"test.mtgproxies"
-    open_database(save_file_path, f"document-v{source_version}", DocumentLoader.MIN_SUPPORTED_SQLITE_VERSION).close()
+    open_database(save_file_path, f"document-v{source_version}").close()
     return save_file_path
 
 
@@ -148,16 +146,14 @@ def _validate_database_schema(db_path: Path):
     :raises AssertionError: If the provided file contains an invalid schema
     :returns: Database schema version
     """
-    target_schema_version = 6
-    db_unsafe = open_database(
-        db_path, f"document-v{target_schema_version}", DocumentLoader.MIN_SUPPORTED_SQLITE_VERSION)
+    target_schema_version = 7
+    db_unsafe = open_database(db_path, f"document-v{target_schema_version}")
     assert_that(
         db_unsafe.execute("PRAGMA application_id").fetchone(), contains_exactly(41325044),
         "Not an MTGProxyPrinter save file!"
     )
     assert_that(db_unsafe.execute("PRAGMA user_version").fetchone(), contains_exactly(target_schema_version))
-    db_known_good = create_in_memory_database(
-        f"document-v{target_schema_version}", DocumentLoader.MIN_SUPPORTED_SQLITE_VERSION)
+    db_known_good = create_in_memory_database(f"document-v{target_schema_version}")
     tables_and_views_query = textwrap.dedent("""\
         SELECT   s.type, s.name,
                  p.cid AS column_id, p.name AS column_name, p.type AS column_type,
@@ -194,9 +190,14 @@ def _validate_database_schema(db_path: Path):
 
 def _validate_saved_document_settings(document: Document, save_file: Path):
     layout = document.page_layout
-    with open_database(save_file, "document-v6", DocumentLoader.MIN_SUPPORTED_SQLITE_VERSION) as save:
+    with open_database(save_file, "document-v7") as save:
         assert_that(
-            save.execute("SELECT COUNT(*) FROM DocumentSettings").fetchone(),
+            save.execute(textwrap.dedent("""
+            SELECT sum(cnt) FROM (
+              SELECT COUNT(1) AS cnt FROM DocumentSettings
+              UNION ALL 
+              SELECT COUNT(1) AS cnt FROM DocumentDimensions
+            )""")).fetchone(),
             contains_exactly(len(dataclasses.astuple(layout)))
         )
         keys = ", ".join(map("'{}'".format, layout.__annotations__.keys()))
@@ -206,22 +207,32 @@ def _validate_saved_document_settings(document: Document, save_file: Path):
               WHERE key IN ({keys})
               ORDER BY key ASC
             """)
-        # TODO: Use the unit-aware matchers!
         assert_that(
             [value for value, in save.execute(query).fetchall()],
             contains_exactly(
-                close_to_(layout.card_bleed.magnitude),
-                close_to_(layout.column_spacing.magnitude),
                 layout.document_name,
-                int(layout.draw_cut_markers),
-                int(layout.draw_sharp_corners),
-                int(layout.draw_page_numbers),
-                close_to_(layout.margin_bottom.magnitude),
-                close_to_(layout.margin_left.magnitude),
-                close_to_(layout.margin_right.magnitude),
-                close_to_(layout.margin_top.magnitude),
-                close_to_(layout.page_height.magnitude),
-                close_to_(layout.page_width.magnitude),
-                close_to_(layout.row_spacing.magnitude),
+                str(layout.draw_cut_markers),
+                str(layout.draw_page_numbers),
+                str(layout.draw_sharp_corners),
+            )
+        )
+        query = textwrap.dedent(f"""\
+                    SELECT value
+                      FROM DocumentDimensions
+                      WHERE key IN ({keys})
+                      ORDER BY key ASC
+                    """)
+        assert_that(
+            [value for value, in save.execute(query).fetchall()],
+            contains_exactly(
+                quantity_close_to(layout.card_bleed),
+                quantity_close_to(layout.column_spacing),
+                quantity_close_to(layout.margin_bottom),
+                quantity_close_to(layout.margin_left),
+                quantity_close_to(layout.margin_right),
+                quantity_close_to(layout.margin_top),
+                quantity_close_to(layout.page_height),
+                quantity_close_to(layout.page_width),
+                quantity_close_to(layout.row_spacing),
             )
         )